:hammer: chore: refactor eslint
