+++ .github/workflows/lint.yml
@@ -13,11 +13,11 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - name: Use Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v1
+        uses: actions/setup-node@v2
         with:
           node-version: ${{ matrix.node-version }}
 
       - name: Install dependencies
         run: npm install
       - name: Run lint
-        run: "npm run lint:check"
+        run: npm run lint
