+++ src/components/modal/Modal.tsx
@@ -1,10 +1,10 @@
-import { closeModal } from "@lib/utils";
+import { closeModal } from "lib/utils";
 import * as React from "react";
 import { createPortal } from "react-dom";
 import { usePortal } from "@casper124578/useful/hooks/usePortal";
 import { ModalStyle, ModalHeader, ModalBody, CloseModal, ModalContainer } from "./styles";
 import { useMounted } from "@casper124578/useful/hooks/useMounted";
-import { ModalIds } from "@lib/constants";
+import { ModalIds } from "lib/constants";
 
 interface Props {
   id: ModalIds;
