+++ src/components/modals/ManageNote.tsx
@@ -4,10 +4,10 @@ import { observer } from "mobx-react-lite";
 import Modal from "@components/modal/Modal";
 import { FormGroup, FormLabel, FormSmall } from "@styles/Auth";
 import { Select, Button } from "@styles/Global";
-import { updateNoteById } from "@actions/note";
-import { isTrue } from "@lib/utils";
+import { updateNoteById } from "actions/note";
+import { isTrue } from "lib/utils";
 import useModalEvent from "@hooks/useModalEvent";
-import { ModalIds } from "@lib/constants";
+import { ModalIds } from "lib/constants";
 import SelectCategory from "@components/SelectCategory";
 import { useStore } from "store/StoreProvider";
 import { toast } from "react-toastify";
