+++ src/lib/middlewares.ts
@@ -13,7 +13,7 @@ import cookieParser from "cookie-parser";
 import rateLimit from "express-rate-limit";
 import cors from "cors";
 import helmet from "helmet";
-import UserModel from "@models/User.model";
+import UserModel from "models/User.model";
 import { IRequest } from "types/IRequest";
 import { ALLOWED_METHODS } from "./shared";
