+++ src/pages/app/index.tsx
@@ -1,17 +1,17 @@
-import { verifyAuth } from "@actions/auth";
+import { verifyAuth } from "actions/auth";
 import { GetServerSideProps } from "next";
 import * as React from "react";
 import { useRouter } from "next/router";
 import dynamic from "next/dynamic";
 import { observer } from "mobx-react-lite";
 import { useMounted } from "@casper124578/useful/hooks/useMounted";
-import { getNoteById, getNotes } from "@actions/note";
+import { getNoteById, getNotes } from "actions/note";
 import Sidebar from "@components/sidebar/Sidebar";
-import { getCategories } from "@actions/categories";
+import { getCategories } from "actions/categories";
 import { AppLayout } from "@styles/Global";
 import Main from "@components/main/Main";
-import { openModal } from "@lib/utils";
-import { ModalIds } from "@lib/constants";
+import { openModal } from "lib/utils";
+import { ModalIds } from "lib/constants";
 import Seo from "@components/Seo";
 import PinModal from "@components/modals/PinModal";
 import { useStore } from "store/StoreProvider";
