+++ src/pages/auth/register.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { useRouter } from "next/router";
 import * as React from "react";
-import { authenticate } from "@actions/auth";
+import { authenticate } from "actions/auth";
 import {
   AuthContainer,
   AuthForm,
