+++ src/pages/logout.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { logout } from "@actions/auth";
+import { logout } from "actions/auth";
 import { useRouter } from "next/router";
 import { useStore } from "store/StoreProvider";
 import { observer } from "mobx-react-lite";
