+++ src/pages/share/[id].tsx
@@ -4,8 +4,8 @@ import Link from "next/link";
 import { observer } from "mobx-react-lite";
 import { GetServerSideProps } from "next";
 import { LinkStyle, ShareStyle, ShareTitle, ShareFooter } from "@styles/Share";
-import { getShareById } from "@actions/note";
-import { verifyAuth } from "@actions/auth";
+import { getShareById } from "actions/note";
+import { verifyAuth } from "actions/auth";
 import NotePreview from "@components/note/NotePreview";
 import Loader from "@components/loader/Loader";
 import Seo from "@components/Seo";
