+++ tsconfig.json
@@ -17,11 +17,11 @@
       "@components/*": ["./src/components/*"],
       "@icons/*": ["./src/components/icons/*"],
       "@hooks/*": ["./src/hooks/*"],
-      "types/*": ["./src/interfaces/*"],
-      "@lib/*": ["./src/lib/*"],
-      "@models/*": ["./src/models/*"],
       "@styles/*": ["./src/styles/*"],
-      "@actions/*": ["./src/store/actions/*"],
+      "lib/*": ["./src/lib/*"],
+      "models/*": ["./src/models/*"],
+      "actions/*": ["./src/store/actions/*"],
+      "types/*": ["./src/interfaces/*"],
       "store/*": ["./src/store/*"]
     },
     "lib": ["dom", "dom.iterable", "esnext"],
